Rename Minimap interface
