--- src/main/java/org/terasology/logic/players/IMinimapSystem.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2013 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.terasology.logic.players;
-
-import org.terasology.minimap.overlays.MinimapOverlay;
-
-/**
- * This class represents the connection to the event system and maintains the HUD element.
- */
-public interface IMinimapSystem {
-
-    /**
-     * @param overlay the overlay to add
-     */
-    public void addOverlay(MinimapOverlay overlay);
-
-    /**
-     * @param overlay the overlay to remove
-     */
-    public void removeOverlay(MinimapOverlay overlay);
-}
